requirements.txt
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
textual==6.5.0
rich==14.2.0
python-dotenv==1.2.1
pytz==2025.2